services/inference_service/app/requirements.txt
fastapi[standard]
# uvicorn
cassandra-driver
python-dotenv
cqlsh